Add files via upload
Lots of work on the color scrolling function for the front LEDs. It's working in the OOP paradigm and accepts booleans to run on any of the quarters of the front lighting strip.

Now I need to rewrite frameAdvance so it will accept any variable as a timer instead of being hardcoded to frameNumber. The front led strip object will need multiple timers to manage running different effects on different portions of the strip.